Rename hook filenames to match current standard.
