+++ src/app/components/header/header.test.tsx
@@ -5,7 +5,7 @@ import userEvent from '@testing-library/user-event';
 
 import { Header } from './header';
 import { RecoilRoot } from 'recoil';
-import * as useAuthMock from '../../hooks/useAuth';
+import * as useAuthMock from '../../hooks/use-auth';
 import { User } from '../../types/user';
 
 describe('Header', () => {
