+++ src/app/pages/details/details.tsx
@@ -3,8 +3,8 @@ import { useParams } from 'react-router-dom';
 import { Card } from '@metrostar/comet-uswds';
 import { Spinner } from '@metrostar/comet-extras';
 import { Launch } from '../../types/launch';
-import useApi from '../../hooks/useApi';
-import useAuth from '../../hooks/useAuth';
+import useApi from '../../hooks/use-api';
+import useAuth from '../../hooks/use-auth';
 import ErrorNotification from '../../components/error-notification/error-notification';
 
 export const Details = (): React.ReactElement => {
