+++ src/app/pages/home/home.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, act } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import * as useAuthMock from '../../hooks/useAuth';
+import * as useAuthMock from '../../hooks/use-auth';
 import { Home } from './home';
 import { RecoilRoot } from 'recoil';
 import { User } from '../../types/user';
