+++ src/app/pages/sign-in/sign-in.test.tsx
@@ -5,7 +5,7 @@ import { BrowserRouter } from 'react-router-dom';
 
 import { SignIn } from './sign-in';
 import { RecoilRoot } from 'recoil';
-import * as useAuthMock from '../../hooks/useAuth';
+import * as useAuthMock from '../../hooks/use-auth';
 import { User } from '../../types/user';
 
 describe('SignIn', () => {
